+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/common/operators/aggregate/PartialAggregateOpExec.scala
@@ -4,30 +4,8 @@ import edu.uci.ics.amber.engine.common.InputExhausted
 import edu.uci.ics.amber.engine.common.tuple.amber.TupleLike
 import edu.uci.ics.texera.workflow.common.operators.OperatorExecutor
 import edu.uci.ics.texera.workflow.common.tuple.Tuple
-import edu.uci.ics.texera.workflow.common.tuple.schema.{Attribute, AttributeType, Schema}
 
 import scala.collection.mutable
-import scala.jdk.CollectionConverters.IterableHasAsJava
-
-object PartialAggregateOpExec {
-  def internalAggObjKey(key: Int): String = {
-    s"__internal_aggregate_partial_object_${key}__"
-  }
-
-  def getOutputSchema(
-      inputSchema: Schema,
-      aggFuncs: List[DistributedAggregation[Object]],
-      groupByKeys: List[String]
-  ): Schema = {
-    Schema
-      .newBuilder()
-      // add group by keys
-      .add(groupByKeys.map(k => inputSchema.getAttribute(k)).asJava)
-      // add intermediate internal aggregation objects
-      .add(aggFuncs.indices.map(i => new Attribute(internalAggObjKey(i), AttributeType.ANY)).asJava)
-      .build()
-  }
-}
 
 class PartialAggregateOpExec(
     val aggFuncs: List[DistributedAggregation[Object]],
